moved beans to SolrJ
